fixed typing bug for favorites"
